fix: move swaggerspec change baseUrl
