+++ src/api/index.js
@@ -3,7 +3,7 @@ import cors from "cors";
 import v1Router from "./v1/routes";
 import CustomError from "./utils/customError";
 import errorHandler from "./utils/errorhandler";
-import { swaggerSpec } from "./swagger-spec";
+import { swaggerSpec } from "./utils/swagger-spec";
 
 const swaggerUi = require("swagger-ui-express");
 
@@ -17,16 +17,15 @@ app.use(cors());
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-// base uri response
-app.get("/", (req, res) => {
-  res.status(200).send("Fury!");
-});
+// router for api version 1
+app.use("/v1", v1Router);
 
 // use swagger-Ui-express for your app documentation endpoint
-app.use("/docs", swaggerUi.serve, swaggerUi.setup(swaggerSpec));
-
-// router for api version 1
-app.use("/api/v1", v1Router);
+const swaggerRouter = express.Router();
+swaggerRouter.use("/", swaggerUi.serve);
+swaggerRouter.get("/", swaggerUi.setup(swaggerSpec));
+swaggerRouter.get("/v1", swaggerUi.setup(swaggerSpec));
+app.use(["/v1", "/"], swaggerRouter);
 
 // routes not found go here
 app.all("*", (req, res, next) => {
