+++ src/api/v1/routes/index.js
@@ -1,10 +1,11 @@
 import express from "express";
+import { swaggerSpec } from "../../utils/swagger-spec";
 
 const router = express.Router();
 
-// default response to base URL
-router.get("/", (req, res) => {
-  res.status(200).send("Employee records microservice is up and running!");
+// route for documentation in json
+router.get("/documentation", (req, res) => {
+  res.status(200).json(swaggerSpec);
 });
 
 export default router;
